Issue #157 Clean up and refactoring
